+++ packages/interledger/src/accounts/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './calculate-balance'
-export * from './validate-id'
