--- packages/interledger/src/accounts/utils/validate-id.ts
@@ -1,5 +1,0 @@
-import { validate, version } from 'uuid'
-
-export function validateId(id: string): boolean {
-  return validate(id) && version(id) === 4
-}
