+++ packages/interledger/src/index.ts
@@ -11,6 +11,7 @@ import { AccountsService } from './accounts/service'
 import { createAssetService } from './asset/service'
 import { createBalanceService } from './balance/service'
 import { createCreditService } from './credit/service'
+import { createDepositService } from './deposit/service'
 import { Logger } from './logger/service'
 
 const logger = Logger
@@ -83,13 +84,24 @@ export const start = async (): Promise<void> => {
     balanceService
   })
 
+  const depositService = createDepositService({
+    logger,
+    assetService,
+    balanceService
+  })
+
   const ratesService = createRatesService({
     pricesUrl,
     pricesLifetime,
     logger
   })
 
-  adminApi = await createAdminApi({ accountsService, creditService, logger })
+  adminApi = await createAdminApi({
+    accountsService,
+    creditService,
+    depositService,
+    logger
+  })
 
   connectorApp = await createConnectorService({
     redis,
