+++ packages/interledger/src/shared/errors.ts
@@ -1,3 +1,5 @@
+import { Asset } from '../asset/service'
+
 export class BalanceTransferError extends Error {
   constructor(public code: number) {
     super()
@@ -9,5 +11,23 @@ export class UnknownBalanceError extends Error {
   constructor(accountId: string) {
     super('Balance not found. accountId=' + accountId)
     this.name = 'UnknownBalanceError'
+  }
+}
+
+export class UnknownLiquidityAccountError extends Error {
+  constructor(asset: Asset) {
+    super(
+      'Unknown liquidity account. code=' + asset.code + ' scale=' + asset.scale
+    )
+    this.name = 'UnknownLiquidityAccountError'
+  }
+}
+
+export class UnknownSettlementAccountError extends Error {
+  constructor(asset: Asset) {
+    super(
+      'Unknown settlement account. code=' + asset.code + ' scale=' + asset.scale
+    )
+    this.name = 'UnknownSettlementAccountError'
   }
 }
