+++ packages/interledger/src/shared/utils.ts
@@ -1,5 +1,5 @@
 import { Pojo, raw } from 'objection'
-import { v4 as uuid } from 'uuid'
+import { v4 as uuid, validate, version } from 'uuid'
 
 export function bigIntToDbUuid(id: bigint): Pojo {
   return raw('?::uuid', [id.toString(16).padStart(32, '0')])
@@ -11,4 +11,8 @@ export function uuidToBigInt(id: string): bigint {
 
 export function randomId(): bigint {
   return uuidToBigInt(uuid())
+}
+
+export function validateId(id: string): boolean {
+  return validate(id) && version(id) === 4
 }
